requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
Pillow==8.3.1
pytesseract==0.3.8
transformers==4.11.3
torch==1.9.0
opencv-python-headless==4.5.3.56
Werkzeug==2.0.1
numpy==1.21.2